test(runtime): add integration tests for newly wired generators
11 integration tests covering Phases 24-29: process evolution events,
organizational events, disruption events, counterfactual JE pairs,
red-flag indicators, collusion rings, temporal vendor chains, entity
relationship graph, cross-process links, and industry-specific output.
Also fixes a minor unnecessary-cast clippy warning in enhanced_orchestrator.

+++ crates/datasynth-runtime/src/enhanced_orchestrator.rs
@@ -5564,7 +5564,7 @@ impl EnhancedOrchestrator {
                     .config
                     .relationship_strength
                     .calculation
-                    .recency_half_life_days as u32,
+                    .recency_half_life_days,
             },
             ..Default::default()
         };
